Revert "* fix: Issues with feedback and sharing email not working (#154)"
This reverts commit 4b8eb7de91e17bf7b6899368e511562c52a683d2.
