--- AmahiAnywhere/AmahiAnywhere/Presentation/Settings/SettingsViewController+TableViewDelegate.swift
@@ -1,47 +1,0 @@
-//
-//  SettingsViewController+TableViewDelegate.swift
-//  AmahiAnywhere
-//
-//  Created by Kanyinsola Fapohunda on 03/04/2019.
-//  Copyright © 2019 Amahi. All rights reserved.
-//
-
-import Foundation
-
-extension SettingsViewController {
-
-// MARK: - Table View data source
-    
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-        
-        let section = indexPath.section
-        let row = indexPath.row
-        
-        switch section {
-            
-        case 0:
-            handleSignOut()
-            break
-        case 1:
-            if row == 0 {
-                performSegue(withIdentifier: SegueIdentfiers.connection, sender: nil)
-            } else if row == 1 {
-                performCacheInvalidation()
-            }
-            
-            break
-        case 2:
-            if row == 1 {
-                openAmahiOnAppStore()
-            } else if row == 2 {
-                handleFeedbackByEmail()
-            } else if row == 3 {
-                handleShareByEmail()
-            }
-            break
-        default:
-            break
-        }
-    }
-}
